+++ bricks/ui_main/__brick__/ui/{{slice.snakeCase()}}/views/{{feature.snakeCase()}}_{{slice.snakeCase()}}_view.dart
@@ -1,15 +1,15 @@
 import 'package:flutter/material.dart';
 import '../../../../../generated/{{module.snakeCase()}}_localizations.dart';
 
-class {{feature.pascalCase()}}DetailView extends StatelessWidget {
+class {{feature.pascalCase()}}{{slice.pascalCase()}}View extends StatelessWidget {
   final Widget content;
-  const {{feature.pascalCase()}}DetailView({super.key, required this.content});
+  const {{feature.pascalCase()}}{{slice.pascalCase()}}View({super.key, required this.content});
 
   @override
   Widget build(BuildContext context) {
     final l10n = {{module.pascalCase()}}Localizations.of(context)!;
     return Scaffold(
-      appBar: AppBar(title: Text(l10n.{{feature.camelCase()}}DetailTitle)),
+      appBar: AppBar(title: Text(l10n.{{feature.camelCase()}}{{slice.pascalCase()}}Title)),
       body: content,
     );
   }