+++ bricks/ui_main/__brick__/ui/{{slice.snakeCase()}}/widgets/{{feature.snakeCase()}}_{{slice.snakeCase()}}_skeleton.dart
@@ -1,8 +1,8 @@
 import 'package:app_ui/app_ui.dart';
 import 'package:flutter/material.dart';
 
-class {{feature.pascalCase()}}DetailSkeleton extends StatelessWidget {
-  const {{feature.pascalCase()}}DetailSkeleton({super.key});
+class {{feature.pascalCase()}}{{slice.pascalCase()}}Skeleton extends StatelessWidget {
+  const {{feature.pascalCase()}}{{slice.pascalCase()}}Skeleton({super.key});
 
   @override
   Widget build(BuildContext context) {